version 1.0.4.3

+++ Source/sql/Create — копия.sql
@@ -1,3 +1,6 @@
+/*version - 1.3.2*/
+
+
 create table config_flags(
 id int NOT NULL AUTO_INCREMENT,
 Описание VARCHAR(100),
@@ -5,9 +8,7 @@ flag int NOT NULL DEFAULT 0,
 PRIMARY KEY(id) 
 )ENGINE=InnoDB CHARACTER SET=UTF8;
 insert into config_flags(id,Описание,flag) values 
-(1,'состояние базы данных 	(0 - Ни разу не запускалась )
-							(1 - Запущена
-							(2 - Остановленна',0);
+(1,'состояние базы данных 	(0 - Остановленна 1 - Запущена ',0),
 (2,'день отчетности успеваемости (1 - 28)',1),
 (3,'день недели отчётности пропусков (0-6)',6);
 
@@ -29,14 +30,12 @@ INSERT INTO prefix(word,Описание) VALUES
 
 
 create table времяПропуски(
-id int NOT NULL AUTO_INCREMENT,
-Даты date NOT NULL UNIQUE,
-PRIMARY KEY(id)
+id int NOT NULL DEFAULT 0,
+Даты date NOT NULL UNIQUE
 )ENGINE=InnoDB CHARACTER SET=UTF8;  
 create table времяУроки(
-id int NOT NULL AUTO_INCREMENT,
-Даты date NOT NULL UNIQUE,
-PRIMARY KEY(id)
+id int NOT NULL DEFAULT 0,
+Даты date NOT NULL UNIQUE
 )ENGINE=InnoDB CHARACTER SET=UTF8;
 create table группы(
 num int NOT NULL DEFAULT 0,
@@ -50,30 +49,32 @@ PRIMARY KEY(Наименование)
 )ENGINE=InnoDB CHARACTER SET=UTF8;
 
 DELIMITER |
-CREATE PROCEDURE update_table (IN name VARCHAR(100),
-IN time VARCHAR(100))
+CREATE PROCEDURE update_table (IN name VARCHAR(100) CHARACTER SET utf8,
+IN tim VARCHAR(100) CHARACTER SET utf8)
     NOT DETERMINISTIC
     SQL SECURITY INVOKER
     COMMENT 'Обновит таблицу в ссотвецтвии 
     с текущими датами'
 BEGIN
-	DECLARE i INT DEFAULT (SELECT COUNT(*) FROM
-	information_schema.COLUMNS WHERE TABLE_NAME=name);
-		SET @sql= CONCAT('select count(Даты) from ',
-		time, ');');
+                SET @i=(SELECT COUNT(*) FROM information_schema.COLUMNS WHERE TABLE_NAME=name);
+		SET @len=0;
+		SET @sql= CONCAT('select count(Даты) INTO @len from ',
+                tim,' where Даты<=NOW();');
 		PREPARE getCountrySql FROM @sql;
 		EXECUTE getCountrySql;
 		DEALLOCATE PREPARE getCountrySql;
-	WHILE i<=len DO
-		SET @temp=(select DATE_FORMAT((select Даты from
-		времяПропуски WHERE времяПропуски.id=i),
-		'%d_%m_%Y'));
-		SET @sql= CONCAT('ALTER TABLE ',name,' ADD
-		',@temp,' int NOT NULL DEFAULT 0;');
+        WHILE @i<=@len DO
+                SET @temp=0;
+                SET @sql= CONCAT('select DATE_FORMAT((select Даты from ',
+                tim,' WHERE id=@i),"%Y_%m_%d") INTO @temp;');
+                PREPARE getCountrySql FROM @sql;
+                EXECUTE getCountrySql;
+                DEALLOCATE PREPARE getCountrySql;
+		SET @sql= CONCAT('ALTER TABLE ',name,' ADD ',@temp,' int NOT NULL DEFAULT 0;'); 
 		PREPARE getCountrySql FROM @sql;
 		EXECUTE getCountrySql;
 		DEALLOCATE PREPARE getCountrySql;
-		SET i=i+1;
+                SET @i=@i+1;
 	END WHILE;
 END|
 DELIMITER ;
@@ -98,7 +99,8 @@ BEGIN
 	update группы set  num=(
 	SELECT @number_:= @number_ + 1 FROM
 	(SELECT @number_:= 0) as tbl);	
-	SET @sql= CONCAT('CREATE TABLE P_',name,'( 
+	SET @sql= CONCAT('CREATE TABLE P_',name,'(
+	num INT, 
 	предмет VARCHAR(100) NOT NULL UNIQUE,
 	FOREIGN KEY(предмет) REFERENCES предметы(Наименование)
 	ON UPDATE CASCADE
@@ -138,7 +140,7 @@ DELIMITER ;
 
 DELIMITER |
 CREATE PROCEDURE addPredmet(IN Name VARCHAR(100)CHARACTER SET utf8,
- IN Gr VARCHAR(100) CHARACTER SET utf8)
+ IN Gr VARCHAR(100) CHARACTER SET utf8, IN NUM_UPDATE INT)
 	NOT DETERMINISTIC
     SQL SECURITY INVOKER
     COMMENT 'Добавит предмет группе'
@@ -157,12 +159,21 @@ BEGIN
 	PREPARE getCountrySql FROM @sql;
 	EXECUTE getCountrySql;
 	DEALLOCATE PREPARE getCountrySql;
+	IF(NUM_UPDATE>0) THEN
+		SET @sql= CONCAT('update P_',Gr,' set  num=(
+		SELECT @number_:= @number_ + 1 FROM
+		(SELECT @number_:= 0) as tbl);');
+		PREPARE getCountrySql FROM @sql;
+		EXECUTE getCountrySql;
+		DEALLOCATE PREPARE getCountrySql;
+	END IF;
+	
 	SET @i=1;
 	SET @len=(select count(word) from prefix); 
 	WHILE @i<=@len DO
 		SET @sql= CONCAT('CREATE TABLE ',(select word from
 		prefix where id=@i),Name,Gr,'( 
-		ФИО VARCHAR(100) NOT NULL, 
+                ФИО VARCHAR(100) NOT NULL UNIQUE,
 		FOREIGN KEY(ФИО) REFERENCES ',Gr,'(ФИО)
 		ON UPDATE CASCADE
 		ON DELETE CASCADE
@@ -171,6 +182,14 @@ BEGIN
 		PREPARE getCountrySql FROM @sql;
 		EXECUTE getCountrySql;
 		DEALLOCATE PREPARE getCountrySql;
+
+                SET @sql= CONCAT('INSERT INTO ',(select word from
+                prefix where id=@i),Name,Gr,'(ФИО) (select ФИО from ',Gr,')');
+                PREPARE getCountrySql FROM @sql;
+                EXECUTE getCountrySql;
+                DEALLOCATE PREPARE getCountrySql;
+
+
 		SET @i=@i+1;
 	END WHILE;	
 END|
@@ -184,50 +203,66 @@ CREATE PROCEDURE deleteGlobalPredmet(IN Name VARCHAR(100) CHARACTER SET utf8)
     COMMENT 'Удалит предмет'
 BEGIN
 	SET @i=1;
-	SET @len=(select max(id) from группы);
+	SET @len=(select max(num) from группы);
 	WHILE @i<=@len DO
-		SET @sql= CONCAT('CALL deletePredmet(',
-		(select Имя from группы where num=@i),');');
-		PREPARE getCountrySql FROM @sql;
-		EXECUTE getCountrySql;
-		DEALLOCATE PREPARE getCountrySql;
+		SET @sq= CONCAT('CALL deletePredmet("',Name,'","',
+		(select Имя from группы where num=@i),'",1);');
+		PREPARE getCountrySq FROM @sq;
+		EXECUTE getCountrySq;
+		DEALLOCATE PREPARE getCountrySq;
 		SET @i=@i+1;
 	END WHILE;
-	SET @sql= CONCAT('DELETE FROM предметы 
-	where Наименование=',Name,');');
-	PREPARE getCountrySql FROM @sql;
-	EXECUTE getCountrySql;
-	DEALLOCATE PREPARE getCountrySql;
+	SET @sq= CONCAT('DELETE FROM предметы 
+	where Наименование="',Name,'";');
+	PREPARE getCountrySq FROM @sq;
+	EXECUTE getCountrySq;
+	DEALLOCATE PREPARE getCountrySq;
 	update предметы set  num=(
 	SELECT @number_:= @number_ + 1 FROM
 	(SELECT @number_:= 0) as tbl);			
 END |
 DELIMITER ;
 
-
 DELIMITER |
-
 CREATE PROCEDURE deletePredmet(IN Name VARCHAR(100) CHARACTER SET utf8,
-IN Gr VARCHAR(100) CHARACTER SET utf8)
-	NOT DETERMINISTIC
+IN Gr VARCHAR(100) CHARACTER SET utf8,IN NUM_UPDATE INT)
+        NOT DETERMINISTIC
     SQL SECURITY INVOKER
-    COMMENT 'Удалит предмет группе'
+    COMMENT 'Удалит предмет группе' 
+
 BEGIN
-	SET @i=1;
-	SET @len=(select count(word) from prefix);
-	WHILE @i<=@len DO
-		SET @sql= CONCAT('DROP TABLE ',(select word from
-		prefix where id=@i),Name,Gr,';');
-		PREPARE getCountrySql FROM @sql;
-		EXECUTE getCountrySql;
-		DEALLOCATE PREPARE getCountrySql;
-		SET @i=@i+1;
-	END WHILE;
-	SET @sql= CONCAT('DELETE FROM P_',Gr,
-	' where предмет=',name);
-		PREPARE getCountrySql FROM @sql;
-		EXECUTE getCountrySql;
-		DEALLOCATE PREPARE getCountrySql;
+	SET @temp=0;
+	SET @sql= CONCAT('select count(предмет) INTO @temp from P_',Gr,';');
+	PREPARE getCountryS FROM @sql;
+	EXECUTE getCountryS;
+	DEALLOCATE PREPARE getCountryS;
+
+	IF(@temp>=1) THEN
+        SET @j=1;
+        SET @le=(select count(word) from prefix);
+        WHILE @j<=@le DO
+                SET @sql= CONCAT('DROP TABLE ',(select word from
+                prefix where id=@j),Name,Gr,';');
+                PREPARE getCountryS FROM @sql;
+                EXECUTE getCountryS;
+                DEALLOCATE PREPARE getCountryS;
+                SET @j=@j+1;
+        END WHILE;
+        SET @sql= CONCAT('DELETE FROM P_',Gr,
+        ' where предмет="',name,'";');
+                PREPARE getCountryS FROM @sql;
+                EXECUTE getCountryS;
+                DEALLOCATE PREPARE getCountryS;
+        IF(NUM_UPDATE>0) THEN
+        	SET @sql= CONCAT('update P_',Gr,' set  num=(
+			SELECT @number_:= @number_ + 1 FROM
+			(SELECT @number_:= 0) as tbl);');
+			PREPARE getCountryS FROM @sql;
+			EXECUTE getCountryS;
+			DEALLOCATE PREPARE getCountryS;
+        END IF;
+END IF;
+
 END|
 DELIMITER ;
 
@@ -270,20 +305,38 @@ CREATE PROCEDURE deleteGroup(IN Name VARCHAR(100) CHARACTER SET utf8)
 	SQL SECURITY INVOKER
 	COMMENT 'удалит группу'
 BEGIN 
-	SET @sql= CONCAT('DROP TABLE ',Name,';');
+	
+	SET @i=1;
+    SET @len=0;
+    SET @sql= CONCAT('select count(предмет) INTO @len from  P_',Name,';');
 	PREPARE getCountrySql FROM @sql;
 	EXECUTE getCountrySql;
 	DEALLOCATE PREPARE getCountrySql;
 	
-	SET @sql= CONCAT('SELECT deletePredmet(предмет) from P_',Name,';');
-	PREPARE getCountrySql FROM @sql;
-	EXECUTE getCountrySql;
-	DEALLOCATE PREPARE getCountrySql;
+    WHILE @i<=@len DO
+    	SET @temp="";
+    	SET @sql= CONCAT('select предмет INTO @temp from P_',Name,' where num=@i;');
+		PREPARE getCountrySql FROM @sql;
+		EXECUTE getCountrySql;
+		DEALLOCATE PREPARE getCountrySql;
+    	
+		SET @sql= CONCAT('call deletePredmet("',@temp,'","',Name,'",0);');
+		PREPARE getCountrySql FROM @sql;
+		EXECUTE getCountrySql;
+		DEALLOCATE PREPARE getCountrySql;
+		SET @i=@i+1;
+	END WHILE;
 	
 	SET @sql= CONCAT('DROP TABLE P_',Name,';');
 	PREPARE getCountrySql FROM @sql;
 	EXECUTE getCountrySql;
 	DEALLOCATE PREPARE getCountrySql;
+	
+	SET @sql= CONCAT('DROP TABLE ',Name,';');
+	PREPARE getCountrySql FROM @sql;
+	EXECUTE getCountrySql;
+	DEALLOCATE PREPARE getCountrySql;
+	
 	
 	SET @sql= CONCAT('DELETE FROM группы WHERE Имя="',Name,'";');
 	PREPARE getCountrySql FROM @sql;
@@ -296,62 +349,12 @@ END|
 DELIMITER ;
 
 
-
-DELIMITER |
-create procedure upgrade_date_sync()
-	NOT DETERMINISTIC
-	SQL SECURITY INVOKER
-	COMMENT 'синхронизирует время '
-BEGIN
-
-	IF((CURDATE()<>(select Даты from времяПропуски where
-	id=(select MAX(id) from времяПропуски ))) AND 
-	(WEEKDAY(NOW())=(select flag from config_flags where id=3))) THEN
-		INSERT INTO времяПропуски (Даты) VALUES (NOW());
-	END IF;
-	
-	IF((CURDATE()<>(select Даты from времяПропуски where
-	id=(select MAX(id) from времяПропуски ))) AND
-	(DAYOFMONTH(NOW())=(select flag from config_flags where id=2))) THEN
-		INSERT INTO времяУроки(Даты) VALUES (NOW());	
-	END IF;
-END|
-DELIMITER ;
-
-DELIMITER |
-create procedure upgrade_date()
-	NOT DETERMINISTIC
-	SQL SECURITY INVOKER
-	COMMENT 'обновит даты базы данных'
-BEGIN
-	IF((select flag from config_flags where id=1)=1) THEN
-		call upgrade_date_sync();
-	END IF;
-END|
-DELIMITER ;
-
-
-DELIMITER |
-create procedure stopBd()
-	NOT DETERMINISTIC
-	SQL SECURITY INVOKER
-	COMMENT 'остоновить базу данных'
-BEGIN
-	IF((select flag from config_flags where id=1)=1) THEN
-		
-		update config_flags set flag=1 
-		where id=2;				
-	END IF;
-END|
-DELIMITER ;
-
 DELIMITER |
 create procedure clear()
 	NOT DETERMINISTIC
 	SQL SECURITY INVOKER
 	COMMENT 'отчистить данные если бд остановлина '
 BEGIN
-	IF((select flag from config_flags where id=1)=2) THEN
 		DROP TABLE времяПропуски;
 		DROP TABLE времяПропуски;
 		
@@ -366,25 +369,6 @@ BEGIN
 		Даты date NOT NULL UNIQUE,
 		PRIMARY KEY(id)
 		)ENGINE=InnoDB CHARACTER SET=UTF8;
-	END IF;
-END|
-DELIMITER ;
-
-DELIMITER |
-create procedure startBd()
-	NOT DETERMINISTIC
-	SQL SECURITY INVOKER
-	COMMENT 'запустит базу данных'
-BEGIN
-	SET @d=CURDATE();
-        IF(DAYOFMONTH(@d)>28) THEN
-		SET @d=@d-3;
-	END IF;
-	IF((select flag from config_flags where id=1)<>1) THEN
-		INSERT INTO времяПропуски (Даты) VALUES (@d);
-		INSERT INTO времяУроки(Даты) VALUES (@d);
-		update config_flags set flag=1 where id=1;	
-	END IF;
 END|
 DELIMITER ;
 
@@ -394,6 +378,10 @@ create procedure showGroup(IN name VARCHAR(100) CHARACTER SET UTF8)
 	SQL SECURITY INVOKER
 	COMMENT 'просмотр пропусков'
 BEGIN
+        SET @sq= CONCAT('CALL update_table("',name,'","времяПропуски");');
+        PREPARE getCountrySq FROM @sq;
+        EXECUTE getCountrySq;
+        DEALLOCATE PREPARE getCountrySq;
 	SET @sql= CONCAT('SELECT * FROM ',name,';');
 	PREPARE getCountrySql FROM @sql;
 	EXECUTE getCountrySql;
@@ -404,14 +392,14 @@ DELIMITER ;
 
 DELIMITER |
 create procedure showPredmet(IN name VARCHAR(100) CHARACTER SET UTF8,
-IN Gr VARCHAR(100) CHARACTER SET UTF8, IN date_ DATE )
+IN Gr VARCHAR(100) CHARACTER SET UTF8, IN date_ VARCHAR(100) )
 	NOT DETERMINISTIC
 	SQL SECURITY INVOKER
 	COMMENT 'просмотр пропусков'
 BEGIN
-	SET @sql= CONCAT('SELECT ',CAST(date_ AS CHAR),' FROM LC_',name,Gr,'PC_',name,Gr,'KRC_',name,Gr,
-	'RGRC_',name,Gr,'LD_',name,Gr,'KRD_',name,Gr,
-	'RGRD_',name,Gr);
+	SET @sql= CONCAT('SELECT ',date_,' FROM LC_',name,Gr,', PC_',name,Gr,', KRC_',name,Gr,
+	', RGRC_',name,Gr,', LD_',name,Gr,', KRD_',name,Gr,
+	', RGRD_',name,Gr);
 	PREPARE getCountrySql FROM @sql;
 	EXECUTE getCountrySql;
 	DEALLOCATE PREPARE getCountrySql;
@@ -430,23 +418,11 @@ END|
 DELIMITER ;
 
 
-/********************************************** тригеры ***********************************************/
-
-
-/* дамп базы данных (добовление в архив)
-mysqldump -uroot -h82.82.82.82 -p database > database.sql
--u или -–user=... - имя пользователя
--h или --host=... - удаленный хост (для локального хоста можно опустить этот параметр)
--p или --password - запросить пароль
-database - имя базы данных
-database.sql - файл для дампа
-
-
-
-mysql -uroot -h82.82.82.82 -p database < database.sql
-mysql> use database;
-mysql> source database.sql
-
-****/
-
-
+DELIMITER |
+create TRIGGER update_timeip AFTER insert ON времяПропуски
+FOR EACH ROW BEGIN
+	update времяПропуски set  id=(
+	SELECT @number_:= @number_ + 1 FROM
+	(SELECT @number_:= 0) as tbl);
+END|
+DELIMITER ;
